mp4: rewrite stsd parsing and add support for ALACMagicCookieDescriptor.
This follows the specs more closely (stricter and handles more cases) and should make future changes easier.
Reading the ALAC descriptor is needed because I see quite a few files where channels/sample_rate in the AudioSampleEntry are incorrect.
As a bonus this gives us the bitrate for ALAC files.