id3: rework saving code
Remove the special case where saving an empty tag
would delete the id3 tag. This doesn't add much besides
slowing things down due to resizing the file and removing
any padding. If deleting is needed there is ID3.delete()

Use the default header parsing code for replacing an existing
tag so we error out in case we don't understand the tag
to replace.