Add API for specifying the amount of padding to use. See #229
This adds a new padding parameter to ID3.save() and AIFF.save() which
can get passed a callback which gets passed a PaddingInfo
object. The callback can then decide how much padding to use
when saving based on how much would be left and the size
of the file.

This also adds the ability for ID3/AIFF tags to shrink again.
The default implemention will remove padding if the padding
makes up more than ~2.5% of the file.