mp3: Rework mpeg frame parsing. Fixes #233
Instead of jumping to multiple points in the stream only looks for syncs from
the beginning and give up after we read too much or re-synced too often. This
prevents us from detecting a valid frame without a xing header or estimating
the length wrong based on the first frame offset.

Idealy it should give up faster, but there are quite a few files out there
where there is lots of padding between the id3 tag and the sync. I guess
due to removing a embedded image and not updating the id3 size to include the
resulting padding.

Mark streams a sketchy if we can only find two consecutive frames and mark as
non sketchy if we have 4 frames or one of the frames has a valid vbri or xing
header.