FileType, Metadata: File-like object support (loading only). See #1
There are now three ways to pass a file. First pass a file path or
file object as the first arg and mutagen will try to do the right thing.
The other two are passing a file name using the "filename" kwarg
or a file object using the "fileobj" kwarg.

Required methods are read/seek/tell. Only StringIO/BytesIO is tested.