mid3cp: add --merge option. Fixes #277
Instead of replacing the whole tag this will copy over
frames and potentially replace existing frames.

The id3 version of the target will be used. Excluded frames and
frames not supported by the target version will not be copied.

In case the target has no ID3 tag this option has no effect.