OggPage.find_last: Add option to only return the last page finishing a packet
In most cases we need the last page for extracting the absolute granule position,
but in case the last page doesn't contain or finish any packets, the position
is set to -1 to indicate that fact. Add a "finishing" option we returns the last
page still finishing a packet and as a result containing a valid position value.