Add support for filesystems which don't support opening files read/write. Fixes #300
GVFS over fuse does not allow opening files read/write and returns ENOTSUP.
Since it still supports opening files for reading and writing we can load the whole file
in memory, pass it to the tagging code and finally write the whole file back.

While it's an ugly and resource intesitive workaround it was relatively easy to add
now that we support file like objects, so let's give it a try.