mid3v2: Add support for WXXX frames. Fixes #344
This uses the same format as TXXX now. Since the ":" separator
is a bit tricky with URLs as values add some examples to the man
page.

The previous WXXX behaviour where it was storing the repr of the
value list was just working by accident.

Also add some more tests for other URL frames and multiple T/WXXX
arguments.