flac: support files with multiple VORBIS_COMMENT blocks. Fixes #377
While the flac spec states that "There may be only one VORBIS_COMMENT block in a stream"
metaflac and other programs don't fail when they parse such a file and just use the first
block (same for VLC and Totem).

Instead of raising when loading a file with multiple blocks just use the first one for the
tags interface and leave the other blocks as is, including writing it back as is.

The only other change is in delete() where we delete all VORBIS_COMMENT blocks, even the
ones not exposed. This makes sure that reloading a file that had its tags deleted doesn't
exposed the other tag block suddenly.