Fix duplicated albumartistsort key
albumartistsort was already added as a registered Text key
some lines above so adding it again as a TXXX key replaces
the previous TSO2 tag registration.

Since picard uses TSO2 for "album artist sort order" and even
if it's not an official tag it seems to be generally supported,
I think it's better to leave that one.